requirements.txt
# Web Scraping
requests>=2.32.0
beautifulsoup4>=4.13.0
bs4>=0.0.1
lxml>=4.9.0
playwright>=1.40.0
selenium>=4.38.0

# Google APIs
google-generativeai>=0.8.0
google-genai>=1.39.0
google-api-python-client>=2.181.0
google-auth>=2.40.0
google-auth-httplib2>=0.2.0
jsonschema>=4.25.0

# Data Processing
pandas>=2.3.0
geopandas>=1.1.0
shapely>=2.1.0
numpy>=2.3.0
pyogrio>=0.11.0
pyproj>=3.7.0

# Image Processing
Pillow>=11.0.0
imageio>=2.37.0
imageio-ffmpeg>=0.6.0
moviepy>=2.0.0

# MySQL Database
PyMySQL>=1.1.0

# AI/ML Libraries
openai>=1.102.0
sentence-transformers>=5.1.0
transformers>=4.57.0
torch>=2.9.0
tokenizers>=0.22.0
huggingface-hub>=0.36.0
safetensors>=0.6.0

# Data Science
scikit-learn>=1.7.0
scipy>=1.16.0

# Excel Processing
openpyxl>=3.1.5

# Utilities
python-dotenv>=1.2.0
python-dateutil>=2.9.0
pytz>=2025.2
tqdm>=4.67.0
tenacity>=9.1.0
tabulate>=0.8.10
loguru>=0.6.0
pydantic>=2.11.0
PyYAML>=6.0.0

# String Processing
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.27.0
RapidFuzz>=3.14.0

# Web Automation
webdriver-manager>=4.0.0
PySocks>=1.7.1

# Networking
httpx>=0.28.0
httpcore>=1.0.0
urllib3>=2.5.0
certifi>=2025.11.0

# Async Support
anyio>=4.10.0
trio>=0.32.0
websockets>=15.0.0
